[skip ci] Only publish NPM packages from `release*` branches.
With PyPI, packages are published from `develop*` branches to another
namespace. With NPM this is not the case so we would publish twice if
both of the kinds of branches were publishing, leading to build errors.

This is different from the `develop`/`release` (without suffixes)
separation that's done in `runtime-*` repositories.